+++ tests/test_widgets.py
@@ -1,6 +1,9 @@
 from __future__ import unicode_literals
 
 from unittest import TestCase
+
+from markupsafe import Markup
+
 from wtforms.widgets import *
 from wtforms.widgets import html5
 
@@ -21,14 +24,6 @@ class DummyField(object):
     iter_choices = lambda x: iter(x.data)
 
 
-class EscapeHtmlTest(TestCase):
-    def test(self):
-        self.assertEqual(core.escape_html('<i class="bar">foo</i>'), '&lt;i class=&quot;bar&quot;&gt;foo&lt;/i&gt;')
-        self.assertEqual(core.escape_html('<i class="bar">foo</i>', quote=False), '&lt;i class="bar"&gt;foo&lt;/i&gt;')
-        self.assertEqual(core.escape_html(HTMLString('<i class="bar">foo</i>')), '<i class="bar">foo</i>')
-        self.assertEqual(core.escape_html(HTMLString('<i class="bar">foo</i>'), quote=False), '<i class="bar">foo</i>')
-
-
 class HTMLParamsTest(TestCase):
     def test_basic(self):
         self.assertEqual(html_params(foo=9, k='wuuu'), 'foo="9" k="wuuu"')
@@ -46,7 +41,7 @@ class HTMLParamsTest(TestCase):
         self.assertEqual(html_params(aria_foo_bar='foobar'), 'aria-foo-bar="foobar"')
 
     def test_quoting(self):
-        self.assertEqual(html_params(foo='hi&bye"quot'), 'foo="hi&amp;bye&quot;quot"')
+        self.assertEqual(html_params(foo='hi&bye"quot'), 'foo="hi&amp;bye&#34;quot"')
 
 
 class ListWidgetTest(TestCase):
@@ -152,10 +147,10 @@ class SelectTest(TestCase):
         )
         self.assertEqual(
             Select.render_option('bar', '<i class="bar"></i>foo', False),
-            '<option value="bar">&lt;i class="bar"&gt;&lt;/i&gt;foo</option>'
+            '<option value="bar">&lt;i class=&#34;bar&#34;&gt;&lt;/i&gt;foo</option>'
         )
         self.assertEqual(
-            Select.render_option('bar', HTMLString('<i class="bar"></i>foo'), False),
+            Select.render_option('bar', Markup('<i class="bar"></i>foo'), False),
             '<option value="bar"><i class="bar"></i>foo</option>'
         )