+++ docs/conf.py
@@ -48,7 +48,7 @@ master_doc = 'index'
 
 # General substitutions.
 project = 'WTForms'
-copyright = '2010-2014 by Thomas Johansson, James Crasta'
+copyright = '2008 by the WTForms team'
 
 # The default replacements for |version| and |release|, also used in various
 # other places throughout the built documents.
@@ -170,8 +170,7 @@ htmlhelp_basename = 'WTFormsdoc'
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, document class [howto/manual]).
 latex_documents = [
-    ('index', 'WTForms.tex', 'WTForms Documentation',
-    'Thomas Johansson, James Crasta', 'manual'),
+    ('index', 'WTForms.tex', 'WTForms Documentation', 'WTForms team', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of