+++ CHANGES.rst
@@ -15,10 +15,15 @@ Unreleased
 -   HTML5 :class:`~fields.html5.IntegerField` and
     :class:`~fields.html5.RangeInput` don't render the ``step="1"``
     attribute by default. (`#343`_)
+-   The :class:`~widgets.TextArea` widget prepends a ``\r\n`` newline
+    when rendering to account for browsers stripping an initial line for
+    display. This does not affect the value. (`#238`_, `#395`_)
 
+.. _#238: https://github.com/wtforms/wtforms/issues/238
 .. _#239: https://github.com/wtforms/wtforms/issues/239
 .. _#343: https://github.com/wtforms/wtforms/pull/343
 .. _#389: https://github.com/wtforms/wtforms/pull/389
+.. _#395: https://github.com/wtforms/wtforms/pull/395
 
 
 Version 2.2